PEAR-1935 - move echarts wrapper, add aria labels
